requirements.txt
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
requests==2.24.0
imgkit==1.0.2
SQLAlchemy==1.3.18